explicitly set content type to 'application/json' for any api json response types
